+++ utils/ghc-toolchain/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Unreleased
 
+* Try the C compiler as a fallback C++ compiler
 * Parse "javascript" and "ghcjs" as a valid Arch and OS
 * First version. See Note [ghc-toolchain overview] in GHC.Toolchain for an overview